+++ graph/arguments.go
@@ -4,6 +4,46 @@ import (
 	"github.com/graphql-go/graphql"
 )
 
+func getIntOrDefault(raw any, defaultValue int) int {
+	val, ok := raw.(int)
+	if !ok {
+		return defaultValue
+	}
+
+	return val
+}
+
+func getIntPointerOrDefault(raw any, defaultValue *int) *int {
+	intVal, ok := raw.(int)
+	if !ok {
+		ptrInt, ok := raw.(*int)
+		if !ok {
+			return defaultValue
+		}
+		return ptrInt
+	}
+
+	return &intVal
+}
+
+func getStringOrDefault(raw any, defaultValue string) string {
+	val, ok := raw.(string)
+	if !ok {
+		return defaultValue
+	}
+
+	return val
+}
+
+func getBoolOrDefault(raw any, defaultValue bool) bool {
+	val, ok := raw.(bool)
+	if !ok {
+		return defaultValue
+	}
+
+	return val
+}
+
 var (
 	LabelInputConfig = graphql.InputObjectFieldConfig{
 		Type:         graphql.String,
