+++ graph/constants.go
@@ -1,0 +1,6 @@
+package graph
+
+const (
+	DEFAULT_LIMIT_VALUE int = 10
+	DEFAULT_PAGE_VALUE  int = 0
+)
